Move screen sizing code
